requirements.txt
# Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.50.0

# Vector DB
pinecone-client>=5.0.0

# Embeddings & Reranking
sentence-transformers>=3.0.0

# Tools
requests>=2.31.0
pypdf>=4.0.0
googletrans==4.0.0-rc.1
apscheduler>=3.10.0
tavily-python>=0.5.0

# Google APIs
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
